+++ tests/aiohttp/app.py
@@ -26,11 +26,11 @@ class MyGraphQLView(GraphQLView):
     graphql_transport_ws_handler_class = DebuggableGraphQLTransportWSHandler
     graphql_ws_handler_class = DebuggableGraphQLWSHandler
 
-    async def get_root_value(self, request: web.Request):
+    async def get_root_value(self, request: web.Request) -> Query:
         return Query()
 
 
-def create_app(**kwargs):
+def create_app(**kwargs) -> web.Application:
     app = web.Application()
     app.router.add_route("*", "/graphql", MyGraphQLView(schema=schema, **kwargs))